backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-allauth==65.15.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
